+++ examples/resources/iosxe_interface_port_channel_subinterface/resource.tf
@@ -6,4 +6,10 @@ resource "iosxe_interface_port_channel_subinterface" "example" {
   ipv4_address                = "192.0.2.2"
   ipv4_address_mask           = "255.255.255.0"
   encapsulation_dot1q_vlan_id = 666
+  helper_addresses = [
+    {
+      address = "10.10.10.10"
+      global  = false
+    }
+  ]
 }