+++ internal/provider/data_source_iosxe_interface_port_channel_subinterface.go
@@ -65,6 +65,27 @@ func (t dataSourceInterfacePortChannelSubinterfaceType) GetSchema(ctx context.Co
 				Type:                types.Int64Type,
 				Computed:            true,
 			},
+			"helper_addresses": {
+				MarkdownDescription: "Specify a destination address for UDP broadcasts",
+				Computed:            true,
+				Attributes: tfsdk.ListNestedAttributes(map[string]tfsdk.Attribute{
+					"address": {
+						MarkdownDescription: "",
+						Type:                types.StringType,
+						Computed:            true,
+					},
+					"global": {
+						MarkdownDescription: "Helper-address is global",
+						Type:                types.BoolType,
+						Computed:            true,
+					},
+					"vrf": {
+						MarkdownDescription: "VRF name for helper-address (if different from interface VRF)",
+						Type:                types.StringType,
+						Computed:            true,
+					},
+				}, tfsdk.ListNestedAttributesOptions{}),
+			},
 		},
 	}, nil
 }