+++ internal/provider/data_source_iosxe_interface_port_channel_subinterface_test.go
@@ -22,6 +22,8 @@ func TestAccDataSourceIosxeInterfacePortChannelSubinterface(t *testing.T) {
 					resource.TestCheckResourceAttr("data.iosxe_interface_port_channel_subinterface.test", "ipv4_address", "192.0.2.2"),
 					resource.TestCheckResourceAttr("data.iosxe_interface_port_channel_subinterface.test", "ipv4_address_mask", "255.255.255.0"),
 					resource.TestCheckResourceAttr("data.iosxe_interface_port_channel_subinterface.test", "encapsulation_dot1q_vlan_id", "666"),
+					resource.TestCheckResourceAttr("data.iosxe_interface_port_channel_subinterface.test", "helper_addresses.0.address", "10.10.10.10"),
+					resource.TestCheckResourceAttr("data.iosxe_interface_port_channel_subinterface.test", "helper_addresses.0.global", "false"),
 				),
 			},
 		},
@@ -65,6 +67,10 @@ resource "iosxe_interface_port_channel_subinterface" "test" {
   ipv4_address = "192.0.2.2"
   ipv4_address_mask = "255.255.255.0"
   encapsulation_dot1q_vlan_id = 666
+  helper_addresses = [{
+    address = "10.10.10.10"
+    global = false
+  }]
   depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, iosxe_restconf.PreReq2, ]
 }