+++ internal/provider/model_iosxe_interface_port_channel_subinterface.go
@@ -5,6 +5,7 @@ package provider
 import (
 	"fmt"
 	"net/url"
+	"reflect"
 	"regexp"
 	"strconv"
 
@@ -15,15 +16,21 @@ import (
 )
 
 type InterfacePortChannelSubinterface struct {
-	Device                   types.String `tfsdk:"device"`
-	Id                       types.String `tfsdk:"id"`
-	Name                     types.String `tfsdk:"name"`
-	Description              types.String `tfsdk:"description"`
-	Shutdown                 types.Bool   `tfsdk:"shutdown"`
-	VrfForwarding            types.String `tfsdk:"vrf_forwarding"`
-	Ipv4Address              types.String `tfsdk:"ipv4_address"`
-	Ipv4AddressMask          types.String `tfsdk:"ipv4_address_mask"`
-	EncapsulationDot1qVlanId types.Int64  `tfsdk:"encapsulation_dot1q_vlan_id"`
+	Device                   types.String                                      `tfsdk:"device"`
+	Id                       types.String                                      `tfsdk:"id"`
+	Name                     types.String                                      `tfsdk:"name"`
+	Description              types.String                                      `tfsdk:"description"`
+	Shutdown                 types.Bool                                        `tfsdk:"shutdown"`
+	VrfForwarding            types.String                                      `tfsdk:"vrf_forwarding"`
+	Ipv4Address              types.String                                      `tfsdk:"ipv4_address"`
+	Ipv4AddressMask          types.String                                      `tfsdk:"ipv4_address_mask"`
+	EncapsulationDot1qVlanId types.Int64                                       `tfsdk:"encapsulation_dot1q_vlan_id"`
+	HelperAddresses          []InterfacePortChannelSubinterfaceHelperAddresses `tfsdk:"helper_addresses"`
+}
+type InterfacePortChannelSubinterfaceHelperAddresses struct {
+	Address types.String `tfsdk:"address"`
+	Global  types.Bool   `tfsdk:"global"`
+	Vrf     types.String `tfsdk:"vrf"`
 }
 
 func (data InterfacePortChannelSubinterface) getPath() string {
@@ -65,6 +72,22 @@ func (data InterfacePortChannelSubinterface) toBody() string {
 	}
 	if !data.EncapsulationDot1qVlanId.Null && !data.EncapsulationDot1qVlanId.Unknown {
 		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"encapsulation.dot1Q.vlan-id", strconv.FormatInt(data.EncapsulationDot1qVlanId.Value, 10))
+	}
+	if len(data.HelperAddresses) > 0 {
+		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address", []interface{}{})
+		for index, item := range data.HelperAddresses {
+			if !item.Address.Null && !item.Address.Unknown {
+				body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address"+"."+strconv.Itoa(index)+"."+"address", item.Address.Value)
+			}
+			if !item.Global.Null && !item.Global.Unknown {
+				if item.Global.Value {
+					body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address"+"."+strconv.Itoa(index)+"."+"global", map[string]string{})
+				}
+			}
+			if !item.Vrf.Null && !item.Vrf.Unknown {
+				body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address"+"."+strconv.Itoa(index)+"."+"vrf", item.Vrf.Value)
+			}
+		}
 	}
 	return body
 }
@@ -109,6 +132,24 @@ func (data *InterfacePortChannelSubinterface) updateFromBody(res gjson.Result) {
 	} else {
 		data.EncapsulationDot1qVlanId.Null = true
 	}
+	for i := range data.HelperAddresses {
+		key := data.HelperAddresses[i].Address.Value
+		if value := res.Get(fmt.Sprintf("%vip.helper-address.#(address==\"%v\").address", prefix, key)); value.Exists() {
+			data.HelperAddresses[i].Address.Value = value.String()
+		} else {
+			data.HelperAddresses[i].Address.Null = true
+		}
+		if value := res.Get(fmt.Sprintf("%vip.helper-address.#(address==\"%v\").global", prefix, key)); value.Exists() {
+			data.HelperAddresses[i].Global.Value = true
+		} else {
+			data.HelperAddresses[i].Global.Value = false
+		}
+		if value := res.Get(fmt.Sprintf("%vip.helper-address.#(address==\"%v\").vrf", prefix, key)); value.Exists() {
+			data.HelperAddresses[i].Vrf.Value = value.String()
+		} else {
+			data.HelperAddresses[i].Vrf.Null = true
+		}
+	}
 }
 
 func (data *InterfacePortChannelSubinterface) fromBody(res gjson.Result) {
@@ -142,6 +183,26 @@ func (data *InterfacePortChannelSubinterface) fromBody(res gjson.Result) {
 	if value := res.Get(prefix + "encapsulation.dot1Q.vlan-id"); value.Exists() {
 		data.EncapsulationDot1qVlanId.Value = value.Int()
 		data.EncapsulationDot1qVlanId.Null = false
+	}
+	if value := res.Get(prefix + "ip.helper-address"); value.Exists() {
+		data.HelperAddresses = make([]InterfacePortChannelSubinterfaceHelperAddresses, 0)
+		value.ForEach(func(k, v gjson.Result) bool {
+			item := InterfacePortChannelSubinterfaceHelperAddresses{}
+			if cValue := v.Get("address"); cValue.Exists() {
+				item.Address.Value = cValue.String()
+				item.Address.Null = false
+			}
+			if cValue := v.Get("global"); cValue.Exists() {
+				item.Global.Value = true
+				item.Global.Null = false
+			}
+			if cValue := v.Get("vrf"); cValue.Exists() {
+				item.Vrf.Value = cValue.String()
+				item.Vrf.Null = false
+			}
+			data.HelperAddresses = append(data.HelperAddresses, item)
+			return true
+		})
 	}
 }
 
@@ -182,9 +243,37 @@ func (data *InterfacePortChannelSubinterface) setUnknownValues() {
 		data.EncapsulationDot1qVlanId.Unknown = false
 		data.EncapsulationDot1qVlanId.Null = true
 	}
+	for i := range data.HelperAddresses {
+		if data.HelperAddresses[i].Address.Unknown {
+			data.HelperAddresses[i].Address.Unknown = false
+			data.HelperAddresses[i].Address.Null = true
+		}
+		if data.HelperAddresses[i].Global.Unknown {
+			data.HelperAddresses[i].Global.Unknown = false
+			data.HelperAddresses[i].Global.Null = true
+		}
+		if data.HelperAddresses[i].Vrf.Unknown {
+			data.HelperAddresses[i].Vrf.Unknown = false
+			data.HelperAddresses[i].Vrf.Null = true
+		}
+	}
 }
 
 func (data *InterfacePortChannelSubinterface) getDeletedListItems(state InterfacePortChannelSubinterface) []string {
 	deletedListItems := make([]string, 0)
+	for _, i := range state.HelperAddresses {
+		if reflect.ValueOf(i.Address.Value).IsZero() {
+			continue
+		}
+		found := false
+		for _, j := range data.HelperAddresses {
+			if i.Address.Value == j.Address.Value {
+				found = true
+			}
+		}
+		if !found {
+			deletedListItems = append(deletedListItems, fmt.Sprintf("%v/ip/helper-address=%v", state.getPath(), i.Address.Value))
+		}
+	}
 	return deletedListItems
 }