+++ internal/provider/model_iosxe_interface_vlan.go
@@ -5,6 +5,7 @@ package provider
 import (
 	"fmt"
 	"net/url"
+	"reflect"
 	"regexp"
 	"strconv"
 
@@ -15,16 +16,23 @@ import (
 )
 
 type InterfaceVLAN struct {
-	Device          types.String `tfsdk:"device"`
-	Id              types.String `tfsdk:"id"`
-	Name            types.Int64  `tfsdk:"name"`
-	Autostate       types.Bool   `tfsdk:"autostate"`
-	Description     types.String `tfsdk:"description"`
-	Shutdown        types.Bool   `tfsdk:"shutdown"`
-	VrfForwarding   types.String `tfsdk:"vrf_forwarding"`
-	Ipv4Address     types.String `tfsdk:"ipv4_address"`
-	Ipv4AddressMask types.String `tfsdk:"ipv4_address_mask"`
-	Unnumbered      types.String `tfsdk:"unnumbered"`
+	Device                     types.String                   `tfsdk:"device"`
+	Id                         types.String                   `tfsdk:"id"`
+	Name                       types.Int64                    `tfsdk:"name"`
+	Autostate                  types.Bool                     `tfsdk:"autostate"`
+	Description                types.String                   `tfsdk:"description"`
+	Shutdown                   types.Bool                     `tfsdk:"shutdown"`
+	VrfForwarding              types.String                   `tfsdk:"vrf_forwarding"`
+	Ipv4Address                types.String                   `tfsdk:"ipv4_address"`
+	Ipv4AddressMask            types.String                   `tfsdk:"ipv4_address_mask"`
+	Unnumbered                 types.String                   `tfsdk:"unnumbered"`
+	IpDhcpRelaySourceInterface types.String                   `tfsdk:"ip_dhcp_relay_source_interface"`
+	HelperAddresses            []InterfaceVLANHelperAddresses `tfsdk:"helper_addresses"`
+}
+type InterfaceVLANHelperAddresses struct {
+	Address types.String `tfsdk:"address"`
+	Global  types.Bool   `tfsdk:"global"`
+	Vrf     types.String `tfsdk:"vrf"`
 }
 
 func (data InterfaceVLAN) getPath() string {
@@ -69,6 +77,25 @@ func (data InterfaceVLAN) toBody() string {
 	}
 	if !data.Unnumbered.Null && !data.Unnumbered.Unknown {
 		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.unnumbered", data.Unnumbered.Value)
+	}
+	if !data.IpDhcpRelaySourceInterface.Null && !data.IpDhcpRelaySourceInterface.Unknown {
+		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.dhcp.Cisco-IOS-XE-dhcp:relay.source-interface", data.IpDhcpRelaySourceInterface.Value)
+	}
+	if len(data.HelperAddresses) > 0 {
+		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address", []interface{}{})
+		for index, item := range data.HelperAddresses {
+			if !item.Address.Null && !item.Address.Unknown {
+				body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address"+"."+strconv.Itoa(index)+"."+"address", item.Address.Value)
+			}
+			if !item.Global.Null && !item.Global.Unknown {
+				if item.Global.Value {
+					body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address"+"."+strconv.Itoa(index)+"."+"global", map[string]string{})
+				}
+			}
+			if !item.Vrf.Null && !item.Vrf.Unknown {
+				body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"ip.helper-address"+"."+strconv.Itoa(index)+"."+"vrf", item.Vrf.Value)
+			}
+		}
 	}
 	return body
 }
@@ -118,6 +145,29 @@ func (data *InterfaceVLAN) updateFromBody(res gjson.Result) {
 	} else {
 		data.Unnumbered.Null = true
 	}
+	if value := res.Get(prefix + "ip.dhcp.Cisco-IOS-XE-dhcp:relay.source-interface"); value.Exists() {
+		data.IpDhcpRelaySourceInterface.Value = value.String()
+	} else {
+		data.IpDhcpRelaySourceInterface.Null = true
+	}
+	for i := range data.HelperAddresses {
+		key := data.HelperAddresses[i].Address.Value
+		if value := res.Get(fmt.Sprintf("%vip.helper-address.#(address==\"%v\").address", prefix, key)); value.Exists() {
+			data.HelperAddresses[i].Address.Value = value.String()
+		} else {
+			data.HelperAddresses[i].Address.Null = true
+		}
+		if value := res.Get(fmt.Sprintf("%vip.helper-address.#(address==\"%v\").global", prefix, key)); value.Exists() {
+			data.HelperAddresses[i].Global.Value = true
+		} else {
+			data.HelperAddresses[i].Global.Value = false
+		}
+		if value := res.Get(fmt.Sprintf("%vip.helper-address.#(address==\"%v\").vrf", prefix, key)); value.Exists() {
+			data.HelperAddresses[i].Vrf.Value = value.String()
+		} else {
+			data.HelperAddresses[i].Vrf.Null = true
+		}
+	}
 }
 
 func (data *InterfaceVLAN) fromBody(res gjson.Result) {
@@ -158,6 +208,30 @@ func (data *InterfaceVLAN) fromBody(res gjson.Result) {
 	if value := res.Get(prefix + "ip.unnumbered"); value.Exists() {
 		data.Unnumbered.Value = value.String()
 		data.Unnumbered.Null = false
+	}
+	if value := res.Get(prefix + "ip.dhcp.Cisco-IOS-XE-dhcp:relay.source-interface"); value.Exists() {
+		data.IpDhcpRelaySourceInterface.Value = value.String()
+		data.IpDhcpRelaySourceInterface.Null = false
+	}
+	if value := res.Get(prefix + "ip.helper-address"); value.Exists() {
+		data.HelperAddresses = make([]InterfaceVLANHelperAddresses, 0)
+		value.ForEach(func(k, v gjson.Result) bool {
+			item := InterfaceVLANHelperAddresses{}
+			if cValue := v.Get("address"); cValue.Exists() {
+				item.Address.Value = cValue.String()
+				item.Address.Null = false
+			}
+			if cValue := v.Get("global"); cValue.Exists() {
+				item.Global.Value = true
+				item.Global.Null = false
+			}
+			if cValue := v.Get("vrf"); cValue.Exists() {
+				item.Vrf.Value = cValue.String()
+				item.Vrf.Null = false
+			}
+			data.HelperAddresses = append(data.HelperAddresses, item)
+			return true
+		})
 	}
 }
 
@@ -202,9 +276,41 @@ func (data *InterfaceVLAN) setUnknownValues() {
 		data.Unnumbered.Unknown = false
 		data.Unnumbered.Null = true
 	}
+	if data.IpDhcpRelaySourceInterface.Unknown {
+		data.IpDhcpRelaySourceInterface.Unknown = false
+		data.IpDhcpRelaySourceInterface.Null = true
+	}
+	for i := range data.HelperAddresses {
+		if data.HelperAddresses[i].Address.Unknown {
+			data.HelperAddresses[i].Address.Unknown = false
+			data.HelperAddresses[i].Address.Null = true
+		}
+		if data.HelperAddresses[i].Global.Unknown {
+			data.HelperAddresses[i].Global.Unknown = false
+			data.HelperAddresses[i].Global.Null = true
+		}
+		if data.HelperAddresses[i].Vrf.Unknown {
+			data.HelperAddresses[i].Vrf.Unknown = false
+			data.HelperAddresses[i].Vrf.Null = true
+		}
+	}
 }
 
 func (data *InterfaceVLAN) getDeletedListItems(state InterfaceVLAN) []string {
 	deletedListItems := make([]string, 0)
+	for _, i := range state.HelperAddresses {
+		if reflect.ValueOf(i.Address.Value).IsZero() {
+			continue
+		}
+		found := false
+		for _, j := range data.HelperAddresses {
+			if i.Address.Value == j.Address.Value {
+				found = true
+			}
+		}
+		if !found {
+			deletedListItems = append(deletedListItems, fmt.Sprintf("%v/ip/helper-address=%v", state.getPath(), i.Address.Value))
+		}
+	}
 	return deletedListItems
 }