+++ internal/provider/resource_iosxe_interface_ethernet.go
@@ -144,6 +144,39 @@ func (t resourceInterfaceEthernetType) GetSchema(ctx context.Context) (tfsdk.Sch
 					helpers.StringEnumValidator("active", "auto", "desirable", "on", "passive"),
 				},
 			},
+			"ip_dhcp_relay_source_interface": {
+				MarkdownDescription: helpers.NewAttributeDescription("Set source interface for relayed messages").String,
+				Type:                types.StringType,
+				Optional:            true,
+				Computed:            true,
+			},
+			"helper_addresses": {
+				MarkdownDescription: helpers.NewAttributeDescription("Specify a destination address for UDP broadcasts").String,
+				Optional:            true,
+				Attributes: tfsdk.ListNestedAttributes(map[string]tfsdk.Attribute{
+					"address": {
+						MarkdownDescription: helpers.NewAttributeDescription("").String,
+						Type:                types.StringType,
+						Optional:            true,
+						Computed:            true,
+						Validators: []tfsdk.AttributeValidator{
+							helpers.StringPatternValidator(0, 0, `(([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])(%[\p{N}\p{L}]+)?`),
+						},
+					},
+					"global": {
+						MarkdownDescription: helpers.NewAttributeDescription("Helper-address is global").String,
+						Type:                types.BoolType,
+						Optional:            true,
+						Computed:            true,
+					},
+					"vrf": {
+						MarkdownDescription: helpers.NewAttributeDescription("VRF name for helper-address (if different from interface VRF)").String,
+						Type:                types.StringType,
+						Optional:            true,
+						Computed:            true,
+					},
+				}, tfsdk.ListNestedAttributesOptions{}),
+			},
 		},
 	}, nil
 }