+++ internal/provider/resource_iosxe_interface_port_channel_test.go
@@ -22,6 +22,9 @@ func TestAccIosxeInterfacePortChannel(t *testing.T) {
 					resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "vrf_forwarding", "VRF1"),
 					resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ipv4_address", "192.0.2.1"),
 					resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ipv4_address_mask", "255.255.255.0"),
+					resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "ip_dhcp_relay_source_interface", "Loopback100"),
+					resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "helper_addresses.0.address", "10.10.10.10"),
+					resource.TestCheckResourceAttr("iosxe_interface_port_channel.test", "helper_addresses.0.global", "false"),
 				),
 			},
 			{
@@ -71,6 +74,11 @@ func testAccIosxeInterfacePortChannelConfig_all() string {
 		vrf_forwarding = "VRF1"
 		ipv4_address = "192.0.2.1"
 		ipv4_address_mask = "255.255.255.0"
+		ip_dhcp_relay_source_interface = "Loopback100"
+		helper_addresses = [{
+		address = "10.10.10.10"
+		global = false
+		}]
   		depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, ]
 	}
 	`